refactor(api): rename v1alpha2 to v1
