+++ api/proto/buf.lock
@@ -2,5 +2,5 @@
 version: v2
 deps:
   - name: buf.build/agntcy/oasf
-    commit: cb065f356f364a96bd37e2289846730c
+    commit: 0da4970ade914f7c995addf2a183aaca
     digest: b5:798ac54c805e39bbd99f4ca9e582d937fa855537f888f4ec102f160d5682116ed7e865f1d2f77977f32058472092d06268813113ba964dccbcc0dee21ddf9867
