--- api/proto/routing/v1alpha2/peer.proto
@@ -1,44 +1,0 @@
-// Copyright AGNTCY Contributors (https://github.com/agntcy)
-// SPDX-License-Identifier: Apache-2.0
-
-syntax = "proto3";
-
-package routing.v1alpha2;
-
-message Peer {
-  // ID of a given peer, typically described by a protocol.
-  // For example:
-  //  - SPIFFE:   "spiffe://example.org/service/foo"
-  //  - JWT:      "jwt:sub=alice,iss=https://issuer.example.com"
-  //  - Tor:      "onion:abcdefghijklmno.onion"
-  //  - DID:      "did:example:123456789abcdefghi"
-  //  - IPFS:     "ipfs:QmYwAPJzv5CZsnAzt8auVZRn2E6sD1c4x8pN5o6d5cW4D5"
-  string id = 1;
-
-  // Multiaddrs for a given peer.
-  // For example:
-  // - "/ip4/127.0.0.1/tcp/4001"
-  // - "/ip6/::1/tcp/4001"
-  // - "/dns4/example.com/tcp/443/https"
-  repeated string addrs = 2;
-
-  // Additional metadata about the peer.
-  map<string, string> annotations = 3;
-
-  // Used to signal the sender's connection capabilities to the peer.
-  PeerConnectionType connection = 4;
-}
-
-enum PeerConnectionType {
-  // Sender does not have a connection to peer, and no extra information (default)
-  PEER_CONNECTION_TYPE_NOT_CONNECTED = 0;
-
-  // Sender has a live connection to peer.
-  PEER_CONNECTION_TYPE_CONNECTED = 1;
-
-  // Sender recently connected to peer.
-  PEER_CONNECTION_TYPE_CAN_CONNECT = 2;
-
-  // Sender made strong effort to connect to peer repeatedly but failed.
-  PEER_CONNECTION_TYPE_CANNOT_CONNECT = 3;
-}
