--- api/proto/search/v1alpha2/record_query.proto
@@ -1,42 +1,0 @@
-// Copyright AGNTCY Contributors (https://github.com/agntcy)
-// SPDX-License-Identifier: Apache-2.0
-
-syntax = "proto3";
-
-package search.v1alpha2;
-
-// A query to match the record against during discovery.
-// For example:
-//  { type: RECORD_QUERY_TYPE_SKILL_NAME, value: "Natural Language Processing" }
-//  { type: RECORD_QUERY_TYPE_LOCATOR, value: "docker-image:https://example.com/docker-image" }
-message RecordQuery {
-  // The type of the query to match against.
-  RecordQueryType type = 1;
-
-  // The query value to match against.
-  string value = 2;
-}
-
-// Defines a list of supported record query types.
-enum RecordQueryType {
-  // Unspecified query type.
-  RECORD_QUERY_TYPE_UNSPECIFIED = 0;
-
-  // Query for an agent name.
-  RECORD_QUERY_TYPE_NAME = 1;
-
-  // Query for an agent version.
-  RECORD_QUERY_TYPE_VERSION = 2;
-
-  // Query for a skill ID.
-  RECORD_QUERY_TYPE_SKILL_ID = 3;
-
-  // Query for a skill name.
-  RECORD_QUERY_TYPE_SKILL_NAME = 4;
-
-  // Query for a locator.
-  RECORD_QUERY_TYPE_LOCATOR = 5;
-
-  // Query for an extension.
-  RECORD_QUERY_TYPE_EXTENSION = 6;
-}
