--- api/proto/sign/v1alpha1/sign_service.proto
@@ -1,100 +1,0 @@
-// Copyright AGNTCY Contributors (https://github.com/agntcy)
-// SPDX-License-Identifier: Apache-2.0
-
-syntax = "proto3";
-
-package sign.v1alpha1;
-
-import "objects/v1/agent.proto";
-
-service SignService {
-  // Sign agents using keyless OIDC based provider or using PEM-encoded private key with an optional passphrase
-  rpc Sign(SignRequest) returns (SignResponse);
-
-  // Verify signed agents using keyless OIDC based provider or using PEM-encoded formatted PEM public key encrypted
-  rpc Verify(VerifyRequest) returns (VerifyResponse);
-}
-
-message SignRequest {
-  // Agent to be signed
-  objects.v1.Agent agent = 1;
-
-  // Signing provider to use
-  SignRequestProvider provider = 2;
-}
-
-message SignRequestProvider {
-  oneof provider {
-    SignWithOIDC oidc = 1;
-    SignWithKey key = 2;
-  }
-}
-
-message SignResponse {
-  // Signed agent
-  objects.v1.Agent agent = 1;
-}
-
-message SignWithOIDC {
-  // List of sign options for OIDC
-  message SignOpts {
-    // Fulcio authority access URL (default value: https://fulcio.sigstage.dev)
-    optional string fulcio_url = 1;
-
-    // Rekor validator access URL (default value: https://rekor.sigstage.dev)
-    optional string rekor_url = 2;
-
-    // Timestamp authority access URL (default value: https://timestamp.sigstage.dev/api/v1/timestamp)
-    optional string timestamp_url = 3;
-
-    // OIDC provider access URL (default value: https://oauth2.sigstage.dev/auth)
-    optional string oidc_provider_url = 4;
-  }
-
-  // Token for OIDC provider
-  string id_token = 1;
-
-  // Signing options for OIDC
-  SignOpts options = 2;
-}
-
-message SignWithKey {
-  // Private key used for signing
-  bytes private_key = 1;
-
-  // Password to unlock the private key
-  optional bytes password = 2;
-}
-
-message VerifyRequest {
-  // Signed agent to be verified
-  objects.v1.Agent agent = 1;
-
-  // Verification provider to use
-  VerifyRequestProvider provider = 2;
-}
-
-message VerifyRequestProvider {
-  oneof provider {
-    VerifyWithOIDC oidc = 1;
-    VerifyWithKey key = 2;
-  }
-}
-
-message VerifyWithOIDC {
-  // Expected issuer in the signed agent
-  string expected_issuer = 1;
-
-  // Expected signer in the signed agent
-  string expected_signer = 2;
-}
-
-message VerifyWithKey {
-  // Public key to validate the signed agent
-  bytes public_key = 1;
-}
-
-message VerifyResponse {
-  // The verify process result
-  bool success = 1;
-}
