+++ api/proto/store/v1/object.proto
@@ -1,0 +1,59 @@
+// Copyright AGNTCY Contributors (https://github.com/agntcy)
+// SPDX-License-Identifier: Apache-2.0
+
+syntax = "proto3";
+
+package store.v1;
+
+// Reference to a content-addressable object.
+message ObjectRef {
+  // Globally-unique content identifier (CID) of the object.
+  // Specs: https://github.com/multiformats/cid
+  string cid = 1;
+}
+
+// Object is a generic data structure that can hold
+// arbitrary data. It is used to store and associate
+// objects in a content-addressable store.
+message Object {
+  // Globally-unique content identifier of the object.
+  // Encodes fully-qualified type of the object as part of "codec".
+  // Specs: https://github.com/multiformats/cid
+  string cid = 1;
+
+  // Type of the object.
+  // Can be extracted from CID.
+  ObjectType type = 2;
+
+  // Metadata associated with the object.
+  map<string, string> annotations = 3;
+
+  // Creation timestamp of the object in the RFC3339 format.
+  // Specs: https://www.rfc-editor.org/rfc/rfc3339.html
+  string created_at = 4;
+
+  // Size of the object in bytes.
+  uint64 size = 5;
+
+  // Opaque data held by this object.
+  // Clients can use {type} to handle processing.
+  optional bytes data = 6;
+}
+
+// Defines a list of supported object data types.
+// Some values may be reserved for future use.
+// These types are used as a "codec" in the CID.
+enum ObjectType {
+  OBJECT_TYPE_UNSPECIFIED = 0; // invalid type, should not be used
+
+  // Common Object Types
+  OBJECT_TYPE_RAW = 1;
+
+  // OASF Record Types (reserved)
+  // For example, v1alpha1 JSON, v1alpha2 Protobuf, etc.
+  reserved 1000 to 9999;
+
+  // OASF Object Types (reserved)
+  // For example, signatures, extensions, etc.
+  reserved 10000 to 19999;
+}
